requirements.txt
flask
flask-cors
aiogram
pydantic-settings
sqlalchemy
python-dotenv
python-multipart
gunicorn
requests
jinja2
